+++ packages/components/src/Shared/Modal/ModalByEntityDomain.tsx
@@ -2,20 +2,42 @@ import { computeDomain } from "@utils/computeDomain";
 import type { AllDomains } from "@core";
 import type { ModalProps } from "./";
 import { ModalLightControls } from "./ModalLightControls";
-
-export interface ModalByEntityDomainProps extends Omit<ModalProps, "children"> {
-  entity: `${AllDomains}.${string}`;
+import { ModalClimateControls } from "./ModalClimateControls";
+import type { ModalLightControlsProps } from "./ModalLightControls";
+import type { ModalClimateControlsProps } from "./ModalClimateControls";
+interface ModalPropsByDomain {
+  light: ModalLightControlsProps;
+  climate: ModalClimateControlsProps;
 }
 
-export function ModalByEntityDomain({
+type EntityDomainProps<E extends `${AllDomains}.${string}`> =
+  E extends `${infer Domain}.${string}`
+    ? Domain extends keyof ModalPropsByDomain
+      ? ModalPropsByDomain[Domain]
+      : Omit<ModalProps, "children">
+    : never;
+
+export type ModalByEntityDomainProps<E extends `${AllDomains}.${string}`> =
+  EntityDomainProps<E> & {
+    entity: E;
+  };
+
+export function ModalByEntityDomain<E extends `${AllDomains}.${string}`>({
   entity,
   ...rest
-}: ModalByEntityDomainProps) {
+}: ModalByEntityDomainProps<E>) {
   const domain = computeDomain(entity);
   switch (domain) {
     case "light":
       return (
         <ModalLightControls entity={entity as `light.${string}`} {...rest} />
+      );
+    case "climate":
+      return (
+        <ModalClimateControls
+          entity={entity as `climate.${string}`}
+          {...rest}
+        />
       );
     default:
       return null;