+++ packages/components/src/index.ts
@@ -34,11 +34,19 @@ export type { FabCardProps } from "./Cards/FabCard";
 // SidebarCard
 export { SidebarCard } from "./Cards/SidebarCard";
 export type { SidebarCardProps } from "./Cards/SidebarCard";
+// ClimateControls
+export { ClimateControls } from "./Shared/ClimateControls";
+export type { ClimateControlsProps } from "./Shared/ClimateControls";
+// ClimateCart
+export { ClimateCard } from "./Cards/ClimateCard";
+export type { ClimateCardProps } from "./Cards/ClimateCard";
 // Modal
 export { Modal } from "./Shared/Modal";
 export type { ModalProps } from "./Shared/Modal";
 export { ModalLightControls } from "./Shared/Modal/ModalLightControls";
 export type { ModalLightControlsProps } from "./Shared/Modal/ModalLightControls";
+export { ModalClimateControls } from "./Shared/Modal/ModalClimateControls";
+export type { ModalClimateControlsProps } from "./Shared/Modal/ModalClimateControls";
 export { ModalByEntityDomain } from "./Shared/Modal/ModalByEntityDomain";
 export type { ModalByEntityDomainProps } from "./Shared/Modal/ModalByEntityDomain";
 // ControlSlider