+++ packages/core/src/HassConnect/index.tsx
@@ -1,6 +1,7 @@
 import type { ReactNode } from "react";
 import { useRef } from "react";
 import { HassProvider } from "./Provider";
+import type { HassProviderProps } from "./Provider";
 
 export type HassConnectProps = {
   /** Any react node to render when authenticated */
@@ -9,8 +10,10 @@ export type HassConnectProps = {
   hassUrl: string;
   /** Any react node to render when not authenticated */
   fallback?: ReactNode;
-  /** called once the connection is successful, and only once */
+  /** called once the entity subscription is successful, and only once */
   onReady?: () => void;
+  /** options for the provider */
+  options?: Omit<HassProviderProps, "children" | "hassUrl">;
 };
 /** This component will show the Home Assistant login form you're used to seeing normally when logging into HA, once logged in you shouldn't see this again unless you clear device storage, once authenticated it will render the child components of HassConnect and provide access to the api. */
 export const HassConnect = ({
@@ -18,6 +21,7 @@ export const HassConnect = ({
   hassUrl,
   fallback = null,
   onReady,
+  options = {},
 }: HassConnectProps): JSX.Element => {
   const onReadyCalled = useRef(false);
   if (!hassUrl) {
@@ -26,7 +30,7 @@ export const HassConnect = ({
     );
   }
   return (
-    <HassProvider hassUrl={hassUrl}>
+    <HassProvider hassUrl={hassUrl} {...options}>
       {(ready) => {
         if (ready && onReady && !onReadyCalled.current) {
           onReady();